+++ x500/PkiPmiExternalDataTypes.go
@@ -551,7 +551,7 @@ type ExtendedNetworkAddress = asn1.RawValue
 //	TerminalType ::= INTEGER {
 //	  telex(3), teletex(4), g3-facsimile(5), g4-facsimile(6), ia5-terminal(7),
 //	  videotex(8)}(0..ub-integer-options)
-type TerminalType = int64
+type TerminalType = int
 
 const TerminalType_Telex TerminalType = 3
